+++ iblrig/session_creator.py
@@ -1,17 +1,21 @@
 #!/usr/bin/env python
-# -*- coding:utf-8 -*-
 # @Author: Niccolò Bonacchi
-# @Date: Thursday, March 28th 2019, 7:19:15 pm
+# @Creation_Date: Thursday, March 28th 2019, 7:19:15 pm
+# @Editor: Michele Fabbri
+# @Edit_Date: 2022-02-01
+"""
+Creates sessions, pre-generates stim and ephys sessions
+"""
 import math
 from pathlib import Path
 
 import matplotlib.pyplot as plt
 import numpy as np
-from ibllib.dsp.smooth import rolling_window as smooth
 
 import iblrig.blocks as blocks
 import iblrig.misc as misc
 import iblrig.path_helper as ph
+from iblrig.misc import smooth_rolling_window as smooth
 
 
 # EPHYS CHOICE WORLD