+++ tasks/_iblrig_tasks_ephysChoiceWorld/session_params.py
@@ -1,15 +1,14 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
 # @Author: Niccolò Bonacchi
-# @Date:   2018-02-02 17:19:09
+# @Creation_Date: 2018-02-02 17:19:09
+# @Editor: Michele Fabbri
+# @Edit_Date: 2022-02-01
 import logging
 import os
 import tkinter as tk
 from pathlib import Path
 from sys import platform
 from tkinter import messagebox
-
-from pythonosc import udp_client
 
 import iblrig.adaptive as adaptive
 import iblrig.ambient_sensor as ambient_sensor
@@ -21,6 +20,7 @@ import iblrig.sound as sound
 import iblrig.user_input as user_input
 from iblrig.path_helper import SessionPathCreator
 from iblrig.rotary_encoder import MyRotaryEncoder
+from pythonosc import udp_client
 
 log = logging.getLogger("iblrig")
 
@@ -221,8 +221,6 @@ class SessionParamHandler(object):
             "Please start recording in spikeglx then press OK\n"
             + "Behavior task will run after you start the bonsai workflow"
         )
-        # from ibllib.graphic import popup
-        # popup(title, msg)
         root = tk.Tk()
         root.withdraw()
         messagebox.showinfo(title, msg)
@@ -321,9 +319,10 @@ if __name__ == "__main__":
         calling bonsai
         turning off lights of bpod board
     """
+    import datetime
+
     import iblrig.fake_task_settings as _task_settings
     import iblrig.fake_user_settings as _user_settings
-    import datetime
 
     dt = datetime.datetime.now()
     dt = [