+++ release_notes.md
@@ -1,5 +1,10 @@
 # **Release notes**
 
+## **Release Notes develop**
+
+- Changed default rig environment name from iblenv to iblrig
+- Removed ibllib dependecies
+
 ## **Release Notes 6.5.3**
 
 - Removed botched releases and patched release notes accordingly.